fix: allow .git directory on parent directory
Currently, the check to validate pre-commit-hooks are getting installed
on a git repository only consider the `.git` directory, if the
pre-commit-hooks config lives in another path that is not the root of
the repository, pre-commit-hooks won't install a script.

This change allows flake.nix files in sub-directories to have the
pre-commit-hooks configuration. This is specially common in monorepos.