typos: fix unexpected behaviour
Enable typos.configPath to be of type path. This way, we can
use excludes from .typos.toml when running
`pre-commit run typos --all-files`. Otherwise, the execution
differs from a normal invocation of typos, which is unexpected
and leads to wrong results.

To not break anything, and to be compliant with the existing
API, I modified configPath to be either a Nix path or a string.

The main motivation for this change is a repository on my machine
where pre-commit ignores directory foobar (which is excluded
by .typos.toml) but passes all 65,000 files of foobar as argument
to typos. In these situation, typos goes nuts and often the
system freezes.

So with this change, I prevent that possibly tens of thousands
of files that should not be checked in any case are passed to
typos, which results in a smooth experience.

[0]: https://github.com/cachix/pre-commit-hooks.nix/pull/387#issuecomment-1893600631