feat(treefmt): add options to settings and disable cache by default
The bootstrapping of the cache does not work well with the
parallelization of pre-commit, also it introduces quite a slow-down.

Another option that works is to configure `require_serial = true`. But
with this there was still an initial slowdown when the cache was not in
place. While disabling the cache was fast every time.

This might work different in different setups thus it makes sense to
make it configurable.