make sure that clang-tools is in the same version across all platforms
We noticed when upgrading nixpkgs in nix, that clang-format would
produce different results on macOS vs Linux.
This is not great behaviour especially in CI, because you cannot win.
This is because llvmPackage might be in a different version.
To solve this, we now always depend on the latest version.