bugfix: remove before/after options from config json file
A recent commit to add ordering priority to the hooks introduced new
keys into the pre-commit JSON config file. These keys should not be
there are they are not keys pre-commit is aware of and produce warning
messages. e.g.:

[WARNING] Unexpected key(s) present on local => treefmt: after, before

This commit fixes it by filter out all keys with names `before` or
`after` before writing the configuration to file.