check: set HOME to a tmp dir to avoid polluting the working dir
pre-commit creates cache files in $HOME/.cache/pre-commit.
If $HOME is set to $PWD, we will pollute the working directory, causing
hook failures.

Previously, we would copy the source files into a nested `src` directory.
Now that Nix already provides the source files in the working directory
for us, to avoid another copy, we set $HOME to a temporary directory.